+++ esp-wifi/src/wifi/os_adapter.rs
@@ -1442,15 +1442,15 @@ pub unsafe extern "C" fn get_time(
 ///   None
 ///
 /// *************************************************************************
-#[cfg(feature = "wifi-logs")]
+#[cfg(feature = "binary-logs")]
 pub unsafe extern "C" fn log_write(
     level: u32,
     _tag: *const crate::binary::c_types::c_char,
     format: *const crate::binary::c_types::c_char,
     args: ...
 ) {
-    let args = core::mem::transmute::<core::ffi::VaListImpl<'_>, core::ffi::VaListImpl<'_>>(args);
-    crate::compat::syslog::syslog(level, format as *const u8, args);
+    #[allow(clippy::missing_transmute_annotations)]
+    crate::binary::log::syslog(level, format as *const u8, core::mem::transmute(args));
 }
 
 /// **************************************************************************
@@ -1469,7 +1469,7 @@ pub unsafe extern "C" fn log_write(
 ///   None
 ///
 /// *************************************************************************
-#[cfg(feature = "wifi-logs")]
+#[cfg(feature = "binary-logs")]
 #[allow(improper_ctypes_definitions)]
 pub unsafe extern "C" fn log_writev(
     level: u32,
@@ -1480,8 +1480,7 @@ pub unsafe extern "C" fn log_writev(
     // annotations on transmute here would require different types for RISC-V and
     // Xtensa - so let's allow `missing_transmute_annotations` in this case
     #[allow(clippy::missing_transmute_annotations)]
-    let args = core::mem::transmute(args);
-    crate::compat::syslog::syslog(level, format as *const u8, args);
+    crate::binary::log::syslog(level, format as *const u8, core::mem::transmute(args));
 }
 
 /// **************************************************************************